Update dev dependencies for and builds of custom CSS code
In superproject's CSS dev dependency src_cr-new.less, update the line numbers in the table of contents section of the file's inline documentation. Next, update the WSU-UE---CSS submodule, which contains universal dev dependencies used to build custom CSS code in every DAESA website development project. Finally, rebuild the files containing custom CSS code that is ready to be deployed on the website associated with this project.